+++ crates/ethereum/src/lib.rs
@@ -2,13 +2,22 @@ use std::collections::BTreeMap;
 use std::future::Future;
 use std::time::Duration;
 
-use alloy::eips::{BlockId, BlockNumberOrTag, RpcBlockHash};
-use alloy::primitives::{Address, B256};
-use alloy::providers::{Provider, ProviderBuilder, WsConnect};
-use alloy::rpc::types::Log;
+use alloy::eips::{BlockId, BlockNumberOrTag};
+use alloy::primitives::Address;
+use alloy::providers::{Provider, ProviderBuilder, RootProvider, WsConnect};
+use alloy::pubsub::PubSubFrontend;
+use alloy::rpc::types::{Filter, Log};
 use anyhow::Context;
-use futures::StreamExt;
-use pathfinder_common::{BlockHash, BlockNumber, EthereumChain, L1ToL2MessageLog, StateCommitment};
+use futures::{FutureExt, StreamExt};
+use pathfinder_common::{
+    BlockHash,
+    BlockNumber,
+    EthereumChain,
+    L1BlockNumber,
+    L1ToL2MessageLog,
+    L1TransactionHash,
+    StateCommitment,
+};
 use primitive_types::{H160, H256, U256};
 use reqwest::{IntoUrl, Url};
 use starknet::StarknetCoreContract;
@@ -36,6 +45,29 @@ pub mod core_addr {
         Decoder::Hex.decode(b"4737c0c1B4D5b1A687B42610DdabEE781152359c");
 }
 
+pub mod block_numbers {
+    use super::{BlockNumber, L1BlockNumber};
+    pub mod mainnet {
+        use super::{BlockNumber, L1BlockNumber};
+        /// The first v0.13.2 block number
+        pub const FIRST_V0_13_2_BLOCK: BlockNumber = BlockNumber::new_or_panic(671_813);
+        /// The first L1 block number with a state update corresponding to
+        /// v0.13.2 of Starknet
+        pub const FIRST_L1_BLOCK_STARKNET_V0_13_2: L1BlockNumber =
+            L1BlockNumber::new_or_panic(20_627_771);
+    }
+
+    pub mod sepolia {
+        use super::{BlockNumber, L1BlockNumber};
+        /// The first v0.13.2 block number
+        pub const FIRST_V0_13_2_BLOCK: BlockNumber = BlockNumber::new_or_panic(86_311);
+        /// The first L1 block number with a state update corresponding to
+        /// v0.13.2 of Starknet
+        pub const FIRST_L1_BLOCK_STARKNET_V0_13_2: L1BlockNumber =
+            L1BlockNumber::new_or_panic(6_453_990);
+    }
+}
+
 /// Events that can be emitted by the Ethereum client
 #[derive(Debug)]
 pub enum EthereumEvent {
@@ -49,6 +81,7 @@ pub struct EthereumStateUpdate {
     pub state_root: StateCommitment,
     pub block_number: BlockNumber,
     pub block_hash: BlockHash,
+    pub l1_block_number: Option<L1BlockNumber>,
 }
 
 /// Ethereum API trait
@@ -56,9 +89,22 @@ pub struct EthereumStateUpdate {
 pub trait EthereumApi {
     async fn get_starknet_state(&self, address: &H160) -> anyhow::Result<EthereumStateUpdate>;
     async fn get_chain(&self) -> anyhow::Result<EthereumChain>;
-    async fn listen<F, Fut>(
+    async fn get_message_logs(
+        &self,
+        address: &H160,
+        from_block: L1BlockNumber,
+        to_block: L1BlockNumber,
+    ) -> anyhow::Result<Vec<L1ToL2MessageLog>>;
+    async fn get_state_updates(
+        &self,
+        address: &H160,
+        from_block: L1BlockNumber,
+        to_block: L1BlockNumber,
+    ) -> anyhow::Result<Vec<EthereumStateUpdate>>;
+    async fn sync_and_listen<F, Fut>(
         &mut self,
         address: &H160,
+        from_block: L1BlockNumber,
         poll_interval: Duration,
         callback: F,
     ) -> anyhow::Result<()>
@@ -71,7 +117,7 @@ pub trait EthereumApi {
 #[derive(Clone, Debug)]
 pub struct EthereumClient {
     url: Url,
-    pending_state_updates: BTreeMap<u64, EthereumStateUpdate>,
+    pending_state_updates: BTreeMap<L1BlockNumber, EthereumStateUpdate>,
 }
 
 impl EthereumClient {
@@ -91,20 +137,16 @@ impl EthereumClient {
         Self::new(url)
     }
 
-    /// Returns the hash of the last finalized block
-    async fn get_finalized_block_hash(&self) -> anyhow::Result<H256> {
+    /// Returns the block number of the last finalized block
+    async fn get_finalized_block_number(&self) -> anyhow::Result<L1BlockNumber> {
         // Create a WebSocket connection
         let ws = WsConnect::new(self.url.clone());
         let provider = ProviderBuilder::new().on_ws(ws).await?;
-
-        // Fetch the finalized block hash
+        // Fetch the finalized block number
         provider
             .get_block_by_number(BlockNumberOrTag::Finalized, false)
             .await?
-            .map(|block| {
-                let block_hash: [u8; 32] = block.header.hash.into();
-                H256::from(block_hash)
-            })
+            .map(|block| L1BlockNumber::new_or_panic(block.header.number))
             .context("Failed to fetch finalized block hash")
     }
 }
@@ -114,9 +156,10 @@ impl EthereumApi for EthereumClient {
     /// Listens for Ethereum events and notifies the caller using the provided
     /// callback. State updates will only be emitted once they belong to a
     /// finalized block.
-    async fn listen<F, Fut>(
+    async fn sync_and_listen<F, Fut>(
         &mut self,
         address: &H160,
+        from_block: L1BlockNumber,
         poll_interval: Duration,
         callback: F,
     ) -> anyhow::Result<()>
@@ -128,9 +171,40 @@ impl EthereumApi for EthereumClient {
         let ws = WsConnect::new(self.url.clone());
         let provider = ProviderBuilder::new().on_ws(ws).await?;
 
+        // Fetch the current Starknet state from Ethereum
+        let state_update = self.get_starknet_state(address).await?;
+        let _ = callback(EthereumEvent::StateUpdate(state_update)).await;
+
+        // Sync logs from the last known L1 handler block up to the latest finalized
+        // block
+        let logs = self
+            .get_message_logs(
+                address,
+                from_block,
+                state_update
+                    .l1_block_number
+                    .expect("missing l1 block number"),
+            )
+            .await?;
+
+        tracing::trace!(
+            number_of_logs=%logs.len(),
+            from_block=%from_block,
+            to_block=?state_update.l1_block_number.unwrap(),
+            "Fetched L1 to L2 message logs",
+        );
+
+        for log in logs {
+            let _ = callback(EthereumEvent::MessageLog(log)).await;
+        }
+
+        // Prevent a gap between the latest confirmed L1 block and the tip of the chain
+        // when syncing logs.
+        let mut logs_in_sync = false;
+
         // Create the StarknetCoreContract instance
-        let address = Address::new((*address).into());
-        let core_contract = StarknetCoreContract::new(address, provider.clone());
+        let core_address = Address::new((*address).into());
+        let core_contract = StarknetCoreContract::new(core_address, provider.clone());
 
         // Listen for L1 to L2 message events
         let mut logs = provider
@@ -147,7 +221,7 @@ impl EthereumApi for EthereumClient {
         // Poll regularly for finalized block number
         let provider_clone = provider.clone();
         let (finalized_block_tx, mut finalized_block_rx) =
-            tokio::sync::mpsc::channel::<BlockNumber>(1);
+            tokio::sync::mpsc::channel::<L1BlockNumber>(1);
         tokio::spawn(async move {
             let mut interval = tokio::time::interval(poll_interval);
             loop {
@@ -156,18 +230,20 @@ impl EthereumApi for EthereumClient {
                     .get_block_by_number(BlockNumberOrTag::Finalized, false)
                     .await
                 {
-                    let block_number = BlockNumber::new_or_panic(finalized_block.header.number);
+                    let block_number = L1BlockNumber::new_or_panic(finalized_block.header.number);
                     let _ = finalized_block_tx.send(block_number).await.unwrap();
                 }
             }
         });
 
-        // Add the finalized block stream to the select! macro
+        // Process incoming events
         loop {
             select! {
                 Some(state_update) = state_updates.next() => {
                     // Decode the state update
-                    let eth_block = state_update.block_number.expect("missing eth block number");
+                    let eth_block = L1BlockNumber::new_or_panic(
+                        state_update.block_number.expect("missing eth block number")
+                    );
                     let state_update: Log<StarknetCoreContract::LogStateUpdate> = state_update.log_decode()?;
                     let block_number = get_block_number(state_update.inner.blockNumber);
                     // Add or remove to/from pending state updates accordingly
@@ -176,6 +252,7 @@ impl EthereumApi for EthereumClient {
                             block_number,
                             block_hash: get_block_hash(state_update.inner.blockHash),
                             state_root: get_state_root(state_update.inner.globalRoot),
+                            l1_block_number: None,
                         };
                         self.pending_state_updates.insert(eth_block, state_update);
                     } else {
@@ -183,12 +260,38 @@ impl EthereumApi for EthereumClient {
                     }
                 }
                 Some(log) = logs.next() => {
+                    // Fetch potentially unsynced logs from the last finalized block number
+                    // up to the block number of the current log.
+                    if !logs_in_sync {
+                        let log_l1_block_number = log.block_number.map(L1BlockNumber::new_or_panic);
+                        let unsynced_logs = self.get_message_logs(
+                            address,
+                            state_update
+                                .l1_block_number
+                                .expect("missing l1 block number"),
+                            log_l1_block_number
+                                .expect("missing l1 block number"),
+                        )
+                        .await?;
+                        tracing::trace!(
+                            number_of_logs=%unsynced_logs.len(),
+                            from_block=?state_update.l1_block_number,
+                            to_block=?log_l1_block_number,
+                            "Fetched unsynced L1 to L2 message logs",
+                        );
+                        for log in unsynced_logs {
+                            callback(EthereumEvent::MessageLog(log)).await;
+                        }
+                        logs_in_sync = true;
+                    }
                     // Decode the message
                     let log: Log<StarknetCoreContract::LogMessageToL2> = log.log_decode()?;
                     // Create L1ToL2MessageHash from the log data
                     let msg = L1ToL2MessageLog {
                         message_hash: H256::from(log.inner.message_hash().to_be_bytes()),
-                        l1_tx_hash: log.transaction_hash.map(|hash| H256::from(hash.0)).unwrap_or_default(),
+                        l1_block_number: log.block_number.map(L1BlockNumber::new_or_panic),
+                        l1_tx_hash: log.transaction_hash.map(|hash| L1TransactionHash::from(hash.0)),
+                        l2_tx_hash: None,
                     };
                     // Emit the message log
                     callback(EthereumEvent::MessageLog(msg)).await;
@@ -196,11 +299,11 @@ impl EthereumApi for EthereumClient {
                 Some(block_number) = finalized_block_rx.recv() => {
                     // Collect all state updates up to (and including) the finalized block
                     let pending_state_updates: Vec<EthereumStateUpdate> = self.pending_state_updates
-                        .range(..=block_number.get())
+                        .range(..=block_number)
                         .map(|(_, &update)| update)
                         .collect();
                     // Remove emitted updates from the map
-                    self.pending_state_updates.retain(|&k, _| k > block_number.get());
+                    self.pending_state_updates.retain(|&k, _| k > block_number);
                     // Emit the state updates
                     for state_update in pending_state_updates {
                         callback(EthereumEvent::StateUpdate(state_update)).await;
@@ -221,9 +324,8 @@ impl EthereumApi for EthereumClient {
         let contract = StarknetCoreContract::new(address, provider);
 
         // Get the finalized block hash
-        let finalized_block_hash = self.get_finalized_block_hash().await?;
-        let block_hash = B256::from(finalized_block_hash.0);
-        let block_id = BlockId::Hash(RpcBlockHash::from_hash(block_hash, None));
+        let finalized_block_number = self.get_finalized_block_number().await?;
+        let block_id = BlockId::Number(BlockNumberOrTag::Number(finalized_block_number.get()));
 
         // Call the contract methods
         let state_root = contract.stateRoot().block(block_id).call().await?;
@@ -235,6 +337,7 @@ impl EthereumApi for EthereumClient {
             state_root: get_state_root(state_root._0),
             block_hash: get_block_hash(block_hash._0),
             block_number: get_block_number(block_number._0),
+            l1_block_number: Some(finalized_block_number),
         })
     }
 
@@ -255,4 +358,204 @@ impl EthereumApi for EthereumClient {
             x => EthereumChain::Other(x),
         })
     }
+
+    /// Get the L1 to L2 message logs for a given address and block range
+    async fn get_message_logs(
+        &self,
+        address: &H160,
+        from_block: L1BlockNumber,
+        to_block: L1BlockNumber,
+    ) -> anyhow::Result<Vec<L1ToL2MessageLog>> {
+        // Create a WebSocket connection
+        let ws = WsConnect::new(self.url.clone());
+        let provider = ProviderBuilder::new().on_ws(ws).await?;
+
+        // Create the StarknetCoreContract instance
+        let address = Address::new((*address).into());
+        let core_contract = StarknetCoreContract::new(address, provider.clone());
+
+        // Create the filter
+        let filter = core_contract.LogMessageToL2_filter().filter;
+
+        // Fetch the logs
+        let mut logs = Vec::new();
+        get_logs_recursive(&provider, &filter, from_block, to_block, &mut logs, 10_000).await?;
+
+        tracing::trace!(
+            number_of_logs=%logs.len(),
+            %from_block,
+            %to_block,
+            "Fetched L1ToL2MessageLog logs"
+        );
+
+        let logs: Vec<L1ToL2MessageLog> = logs
+            .into_iter()
+            .filter_map(|log| {
+                log.log_decode::<StarknetCoreContract::LogMessageToL2>()
+                    .ok()
+                    .map(|decoded| L1ToL2MessageLog {
+                        message_hash: H256::from(decoded.inner.message_hash().to_be_bytes()),
+                        l1_block_number: log.block_number.map(L1BlockNumber::new_or_panic),
+                        l1_tx_hash: log
+                            .transaction_hash
+                            .map(|hash| L1TransactionHash::from(hash.0)),
+                        l2_tx_hash: None,
+                    })
+            })
+            .collect();
+
+        Ok(logs)
+    }
+
+    /// Get state updates for a given address and block range
+    async fn get_state_updates(
+        &self,
+        address: &H160,
+        from_block: L1BlockNumber,
+        to_block: L1BlockNumber,
+    ) -> anyhow::Result<Vec<EthereumStateUpdate>> {
+        // Create a WebSocket connection
+        let ws = WsConnect::new(self.url.clone());
+        let provider = ProviderBuilder::new().on_ws(ws).await?;
+
+        // Create the StarknetCoreContract instance
+        let address = Address::new((*address).into());
+        let core_contract = StarknetCoreContract::new(address, provider.clone());
+
+        // Create the filter
+        let filter = core_contract.LogStateUpdate_filter().filter;
+
+        // Fetch the logs
+        let mut logs = Vec::new();
+        get_logs_recursive(&provider, &filter, from_block, to_block, &mut logs, 2_000).await?;
+
+        let logs: Vec<EthereumStateUpdate> = logs
+            .into_iter()
+            .filter_map(|log| {
+                log.log_decode::<StarknetCoreContract::LogStateUpdate>()
+                    .ok()
+                    .map(|decoded| EthereumStateUpdate {
+                        state_root: get_state_root(decoded.inner.globalRoot),
+                        block_hash: get_block_hash(decoded.inner.blockHash),
+                        block_number: get_block_number(decoded.inner.blockNumber),
+                        l1_block_number: log.block_number.map(L1BlockNumber::new_or_panic),
+                    })
+            })
+            .collect();
+
+        Ok(logs)
+    }
+}
+
+/// Recursively fetches logs while respecting provider limits
+fn get_logs_recursive<'a>(
+    provider: &'a RootProvider<PubSubFrontend>,
+    base_filter: &'a Filter,
+    from_block: L1BlockNumber,
+    to_block: L1BlockNumber,
+    logs: &'a mut Vec<Log>,
+    // Limits
+    max_block_range: u64,
+) -> futures::future::BoxFuture<'a, anyhow::Result<()>> {
+    async move {
+        // Nothing to do
+        if from_block > to_block {
+            return Ok(());
+        }
+
+        // If the block range exceeds the maximum, split it
+        let block_range = to_block.get() - from_block.get() + 1;
+        if block_range > max_block_range {
+            let mid_block = from_block + block_range / 2;
+            get_logs_recursive(
+                provider,
+                base_filter,
+                from_block,
+                mid_block - 1,
+                logs,
+                max_block_range,
+            )
+            .await?;
+            get_logs_recursive(
+                provider,
+                base_filter,
+                mid_block,
+                to_block,
+                logs,
+                max_block_range,
+            )
+            .await?;
+
+            return Ok(());
+        }
+
+        // Adjust the base filter to the current block range
+        let from_block_id = BlockNumberOrTag::Number(from_block.get());
+        let to_block_id = BlockNumberOrTag::Number(to_block.get());
+        let filter = (*base_filter)
+            .clone()
+            .from_block(from_block_id)
+            .to_block(to_block_id);
+
+        // Attempt to fetch the logs
+        let result = provider.get_logs(&filter).await;
+        match result {
+            Ok(new_logs) => {
+                logs.extend(new_logs);
+            }
+            Err(e) => {
+                tracing::debug!(%from_block, error=?e, "Get logs error at block");
+                if let Some(err) = e.as_error_resp() {
+                    // Retry the request splitting the block range in half
+                    //
+                    // Multiple providers have multiple restrictions. The max range limit can help
+                    // with the obvious restrictions, but not with those that depend on the response
+                    // size. And because there's no way we can predict this, retrying with a smaller
+                    // range is the best we can do.
+                    if err.is_retry_err() {
+                        tracing::debug!(
+                            %from_block,
+                            error=?err,
+                            "Retrying request (splitting) at block"
+                        );
+                        let mid_block = from_block + block_range / 2;
+                        get_logs_recursive(
+                            provider,
+                            base_filter,
+                            from_block,
+                            mid_block - 1,
+                            logs,
+                            max_block_range,
+                        )
+                        .await?;
+                        get_logs_recursive(
+                            provider,
+                            base_filter,
+                            mid_block,
+                            to_block,
+                            logs,
+                            max_block_range,
+                        )
+                        .await?;
+                        return Ok(());
+                    } else {
+                        tracing::error!(
+                            %from_block,
+                            error=?err,
+                            "get_logs provider error"
+                        );
+                    }
+                } else {
+                    tracing::error!(
+                        %from_block,
+                        error=?e,
+                        "get_logs: Unknown error"
+                    );
+                }
+            }
+        }
+
+        Ok(())
+    }
+    .boxed()
 }