+++ crates/pathfinder/src/state/sync.rs
@@ -15,6 +15,7 @@ use pathfinder_common::state_update::{ContractUpdate, SystemContractUpdate};
 use pathfinder_common::{
     BlockCommitmentSignature,
     Chain,
+    L1BlockNumber,
     L1ToL2MessageLog,
     PublicKey,
     ReceiptCommitment,
@@ -104,6 +105,15 @@ where
             chain: value.chain,
             core_address: value.core_address,
             poll_interval: value.l1_poll_interval,
+            last_synced_l1_handler_block: match value.chain {
+                Chain::Mainnet => {
+                    pathfinder_ethereum::block_numbers::mainnet::FIRST_L1_BLOCK_STARKNET_V0_13_2
+                }
+                Chain::SepoliaTestnet | Chain::SepoliaIntegration => {
+                    pathfinder_ethereum::block_numbers::sepolia::FIRST_L1_BLOCK_STARKNET_V0_13_2
+                }
+                _ => L1BlockNumber::GENESIS,
+            },
         }
     }
 }
@@ -178,7 +188,7 @@ where
         ) -> F2
         + Copy,
 {
-    let l1_context = L1SyncContext::from(&context);
+    let mut l1_context = L1SyncContext::from(&context);
     let l2_context = L2SyncContext::from(&context);
 
     let SyncContext {
@@ -209,6 +219,7 @@ where
 
     let (event_sender, event_receiver) = mpsc::channel(8);
 
+    // Get the latest block from the database
     let l2_head = tokio::task::block_in_place(|| -> anyhow::Result<_> {
         let tx = db_conn.transaction()?;
         let l2_head = tx
@@ -219,11 +230,13 @@ where
         Ok(l2_head)
     })?;
 
+    // Get the latest block from the sequencer
     let gateway_latest = sequencer
         .head()
         .await
         .context("Fetching latest block from gateway")?;
 
+    // Keep polling the sequencer for the latest block
     let (tx_latest, rx_latest) = tokio::sync::watch::channel(gateway_latest);
     let mut latest_handle = tokio::spawn(l2::poll_latest(
         sequencer.clone(),
@@ -246,10 +259,30 @@ where
         gossiper,
     ));
 
+    // Fetch the latest known block with an L1 handler tx
+    let l1_handler_block = tokio::task::block_in_place(|| -> anyhow::Result<_> {
+        let tx = db_conn.transaction()?;
+        let l1_handler_block = tx
+            .last_known_l1_block_with_l1_to_l2_message()
+            .context("Fetching latest block with L1 handler tx")?;
+        let default_block_number = match context.chain {
+            Chain::Mainnet => {
+                pathfinder_ethereum::block_numbers::mainnet::FIRST_L1_BLOCK_STARKNET_V0_13_2
+            }
+            Chain::SepoliaTestnet | Chain::SepoliaIntegration => {
+                pathfinder_ethereum::block_numbers::sepolia::FIRST_L1_BLOCK_STARKNET_V0_13_2
+            }
+            _ => L1BlockNumber::GENESIS,
+        };
+        Ok(l1_handler_block.unwrap_or(default_block_number))
+    })?;
+    l1_context.last_synced_l1_handler_block = l1_handler_block;
+
     // Start L1 producer task. Clone the event sender so that the channel remains
     // open even if the producer task fails.
     let mut l1_handle = tokio::spawn(l1_sync(event_sender.clone(), l1_context.clone()));
 
+    // Fetch latest blocks from storage
     let latest_blocks = latest_n_blocks(&mut db_conn, block_cache_size)
         .await
         .context("Fetching latest blocks from storage")?;
@@ -686,8 +719,42 @@ async fn consumer(
                 }
             }
             L1ToL2Message(msg) => {
-                tracing::trace!("Got a new L1 to L2 message log: {:?}", msg);
-                // todo!()
+                tracing::trace!(message_hash=%msg.message_hash, "Got an L1ToL2Message");
+                tokio::task::block_in_place(|| {
+                    // Note: There's always an L1 tx hash and block number (hence the `expect`)
+                    let l1_tx_hash = msg.l1_tx_hash.expect("missing l1 tx hash");
+                    let l1_block_number = msg.l1_block_number.expect("missing l1 block number");
+
+                    let tx = db_conn
+                        .transaction_with_behavior(TransactionBehavior::Immediate)
+                        .context("Creating database transaction")?;
+
+                    // If we have an L2 tx hash for this message, we can associate the L1 handler tx
+                    // with the L2 tx
+                    if let Some(L1ToL2MessageLog {
+                        l2_tx_hash: Some(l2_tx_hash),
+                        ..
+                    }) = tx.fetch_l1_to_l2_message_log(&msg.message_hash)?
+                    {
+                        tracing::trace!(%l1_tx_hash, %l2_tx_hash, "Found L2 tx for L1 tx");
+                        tx.insert_l1_handler_tx(l1_block_number, l1_tx_hash, l2_tx_hash)?;
+                        tx.remove_l1_to_l2_message_log(&msg.message_hash)?;
+                    }
+                    // Otherwise, we insert the message log with an empty L2 tx hash
+                    else {
+                        tracing::trace!(%l1_tx_hash, "L2 tx NOT found for L1 tx");
+                        let msg_log = L1ToL2MessageLog {
+                            message_hash: msg.message_hash,
+                            l1_block_number: Some(l1_block_number),
+                            l1_tx_hash: Some(l1_tx_hash),
+                            l2_tx_hash: None,
+                        };
+                        tx.upsert_l1_to_l2_message_log(&msg_log)?;
+                    }
+                    tx.commit()
+                        .context("Associating L1 and L2 transactions from message log")
+                })
+                .with_context(|| format!("Insert L1 to L2 message log: {:?}", msg))?;
             }
         }
     }